TSI[05012016] Penambahan confirmation dialog dan error exception

+++ han/src/module/dryin/ui/DryInCreatePanel.java
@@ -184,7 +184,7 @@ public class DryInCreatePanel extends JPanel implements Bridging {
 			listOfChamber = ServiceFactory.getDryInBL().getAllChamber();
 		} catch (SQLException e1) {
 			e1.printStackTrace();
-			JOptionPane.showMessageDialog(null, "Terjadi kesalahan pada sistem.", "Error", JOptionPane.ERROR_MESSAGE);
+			DialogBox.showErrorException();
 		}
 
 		cbChamber = new ComboBox<Chamber>();
@@ -234,7 +234,10 @@ public class DryInCreatePanel extends JPanel implements Bridging {
 		btnDeletePicTally = new JButton("Hapus");
 		btnDeletePicTally.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent arg0) {
-				doDeletePicTally();
+				int response = DialogBox.showDeleteChoice();
+				if (response == JOptionPane.YES_OPTION) {
+					doDeletePicTally();
+				}
 			}
 		});
 		btnDeletePicTally.setBounds(925, 200, 100, 30);
@@ -330,7 +333,10 @@ public class DryInCreatePanel extends JPanel implements Bridging {
 		btnInsertDryInPallet = new JButton("Insert");
 		btnInsertDryInPallet.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
-				doInsertDryInPallet();
+				int response = DialogBox.showInsertChoice();
+				if (response == JOptionPane.YES_OPTION) {
+					doInsertDryInPallet();
+				}
 			}
 		});
 		btnInsertDryInPallet.setBounds(220, 540, 100, 30);
